requirements.txt
google-api-python-client==2.155.0
python-dotenv==1.0.1
boto3==1.35.81
pandas==2.2.3
tenacity==9.0.0
youtube-transcript-api==0.6.3